+++ src/app/ApplicationTemplate.Access/Configuration/DadJokesEndpointOptions.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ApplicationTemplate;
-
-public class DadJokesApiClientOptions : ApiClientOptions
-{
-}