+++ src/app/ApplicationTemplate.Shared.Views/Content/Menu.xaml.cs
@@ -1,9 +1,0 @@
-﻿namespace ApplicationTemplate.Views.Content;
-
-public sealed partial class Menu : AttachableUserControl
-{
-	public Menu()
-	{
-		this.InitializeComponent();
-	}
-}