+++ src/app/TradeZeroApp.Access/ApiClients/Posts/ErrorData.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TradeZeroApp.DataAccess;
+
+public class ErrorData
+{
+}