+++ src/app/TradeZeroApp.Access/Configuration/DadJokesEndpointOptions.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TradeZeroApp;
+
+public class DadJokesApiClientOptions : ApiClientOptions
+{
+}