+++ src/app/TradeZeroApp.Presentation/Framework/Connectivity/ConnectivityChangedEventArgs.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace TradeZeroApp;
+
+public sealed class ConnectivityChangedEventArgs : EventArgs
+{
+	public ConnectivityChangedEventArgs(NetworkAccess networkAccess)
+	{
+		NetworkAccess = networkAccess;
+	}
+
+	public NetworkAccess NetworkAccess { get; }
+}