+++ src/app/TradeZeroApp.Presentation/ViewModels/Authentication/SentEmailConfirmationPageViewModel.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace TradeZeroApp.Presentation;
+
+public class SentEmailConfirmationPageViewModel : ViewModel
+{
+}