+++ src/app/TradeZeroApp.Presentation/ViewModels/Diagnostics/NetworkDiagnosticsViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace TradeZeroApp.Presentation;
+
+public class NetworkDiagnosticsViewModel : ViewModel
+{
+	public NetworkDiagnosticsViewModel()
+	{
+	}
+}