+++ src/app/TradeZeroApp.Shared.Views/Content/Authentication/CreateAccountPage.xaml.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.UI.Xaml.Controls;
+
+namespace TradeZeroApp.Views.Content;
+
+public sealed partial class CreateAccountPage : Page
+{
+	public CreateAccountPage()
+	{
+		this.InitializeComponent();
+	}
+}