+++ src/app/TradeZeroApp.Shared.Views/Content/Menu.xaml.cs
@@ -1,0 +1,9 @@
+﻿namespace TradeZeroApp.Views.Content;
+
+public sealed partial class Menu : AttachableUserControl
+{
+	public Menu()
+	{
+		this.InitializeComponent();
+	}
+}