+++ src/app/TradeZeroApp.Tests.Functional/Usings.cs
@@ -1,0 +1,7 @@
+﻿global using System;
+global using TradeZeroApp.Presentation;
+global using Chinook.DynamicMvvm;
+global using Chinook.SectionsNavigation;
+global using Chinook.StackNavigation;
+global using FluentAssertions;
+global using Xunit;